feat(*): hook with data
