+++ app/p/[slug]/contributions/[contributionId]/hooks/use-applications/use-applications.types.ts
@@ -1,0 +1,17 @@
+import { GetApplicationPageResponse } from "api-client/resources/applications/types";
+
+export namespace TUseApplications {
+  type Applications = GetApplicationPageResponse["applications"];
+
+  export interface Props {
+    search: string;
+  }
+
+  export interface Return {
+    newComersApplications?: Applications;
+    projectMembersApplications?: Applications;
+    title: string;
+    newComersApplicationsHasNextPage: boolean;
+    projectMembersApplicationsHasNextPage: boolean;
+  }
+}
